Modify Create function to take index settings
The new Create function will allow for Elasticsearch indices to be
created that have default settings associated with them. Some of these
settings that are likely to be used for our indices include default
mappings, turning off dynamic mappings (so new fields are rejected), and
turning off analysis of certain fields (specifically, UUIDs).